+++ docs/README.md
@@ -1,4 +1,4 @@
-# Yapper User Guide
+# yapper.Yapper User Guide
 
 // Update the title above to match the actual product name
