+++ src/main/java/yapper/command/AddDeadlineCommand.java
@@ -1,3 +1,7 @@
+package yapper.command;
+
+import yapper.*;
+import yapper.task.*;
 import java.io.IOException;
 
 public class AddDeadlineCommand extends Command {
