+++ src/main/java/yapper/command/Command.java
@@ -1,3 +1,9 @@
+package yapper.command;
+
+import yapper.*;
+import yapper.task.TaskList;
+
+
 import java.io.IOException;
 
 public abstract class Command {
