+++ src/main/java/yapper/command/ExitCommand.java
@@ -1,3 +1,8 @@
+package yapper.command;
+
+import yapper.*;
+import yapper.task.*;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
