+++ src/main/java/yapper/command/UnmarkCommand.java
@@ -1,9 +1,13 @@
+package yapper.command;
+
+import yapper.*;
+import yapper.task.*;
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
     private final String input;
 
-    public DeleteCommand(String input) {
+    public UnmarkCommand(String input) {
         this.input = input;
     }
 
@@ -11,8 +15,8 @@ public class DeleteCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
         try {
             int index = Integer.parseInt(input.trim()) - 1;
-            Task removedTask = tasks.deleteTask(index);
-            ui.showMessage("Noted. I've removed this task:\n  " + removedTask + "\nNow you have " + tasks.size() + " tasks in the list.");
+            Task unmarkedTask = tasks.markTaskAsNotDone(index);
+            ui.showMessage("OK, I've marked this task as not done yet:\n  " + unmarkedTask);
             storage.saveTasks(tasks.getTasks());
         } catch (NumberFormatException | IndexOutOfBoundsException e) {
             throw new IllegalArgumentException("Invalid task number.");
