+++ src/main/java/yapper/task/Deadline.java
@@ -1,8 +1,10 @@
+package yapper.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 public class Deadline extends Task {
-    protected LocalDateTime by;
+    public LocalDateTime by;
 
     public Deadline(String description, String by) {
         super(description);
