+++ src/main/java/yapper/task/Event.java
@@ -1,9 +1,11 @@
+package yapper.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 public class Event extends Task {
-    protected LocalDateTime from;
-    protected LocalDateTime to;
+    public LocalDateTime from;
+    public LocalDateTime to;
 
     public Event(String description, String from, String to) {
         super(description);
