+++ src/main/java/yapper/task/Todo.java
@@ -1,3 +1,5 @@
+package yapper.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
