+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Yapper!
+Hello! I'm yapper.Yapper!
 Yapa Yapa Yapa
 ____________________________________________________________
