+++ text-ui-test/runtest.sh
@@ -1,36 +1,31 @@
 #!/usr/bin/env bash
 
 # create bin directory if it doesn't exist
-if [ ! -d "../bin" ]
-then
+if [ ! -d "../bin" ]; then
     mkdir ../bin
 fi
 
 # delete output from previous run
-if [ -e "./ACTUAL.TXT" ]
-then
+if [ -e "./ACTUAL.TXT" ]; then
     rm ACTUAL.TXT
 fi
 
-# ensure the yapper.txt file is emptied (clean slate)
-if [ -e "./data/yapper.txt" ]
-then
-    > ./data/yapper.txt  # Empty the file without deleting it
+# ensure yapper.txt is reset
+if [ -e "./data/yapper.txt" ]; then
+    > ./data/yapper.txt
 else
-    # create the directory and file if they don't exist
     mkdir -p ./data
     touch ./data/yapper.txt
 fi
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/*.java
-then
+if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/**/*.java; then
     echo "********** BUILD FAILURE **********"
     exit 1
 fi
 
 # run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ../bin Yapper < input.txt > ACTUAL.TXT
+java -classpath ../bin yapper.Yapper < input.txt > ACTUAL.TXT
 
 # convert to UNIX format
 cp EXPECTED.TXT EXPECTED-UNIX.TXT
@@ -38,8 +33,7 @@ dos2unix ACTUAL.TXT EXPECTED-UNIX.TXT
 
 # compare the output to the expected output
 diff ACTUAL.TXT EXPECTED-UNIX.TXT
-if [ $? -eq 0 ]
-then
+if [ $? -eq 0 ]; then
     echo "Test result: PASSED"
     exit 0
 else
